% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat.R
\name{osmium_cat}
\alias{osmium_cat}
\title{Concatenate OSM files and convert between formats}
\usage{
osmium_cat(
  input,
  output = NULL,
  output_format = NULL,
  object_type = NULL,
  clean = NULL,
  buffer_data = FALSE,
  input_format = NULL,
  overwrite = FALSE,
  fsync = FALSE,
  generator = NULL,
  output_header = NULL,
  verbose = FALSE
)
}
\arguments{
\item{input}{Character vector of one or more input OSM file paths.}

\item{output}{Output file path. If NULL, output goes to stdout and is
returned as a character string (see Value).}

\item{output_format}{Format of the output file (overrides the
extension of `output`), e.g. "pbf", "xml", "opl".}

\item{object_type}{Character vector restricting output to given
object types: any of "node", "way", "relation", "changeset".}

\item{clean}{Character vector of attributes to strip from output
objects: any of "version", "changeset", "timestamp", "uid", "user".}

\item{buffer_data}{If TRUE, buffer all data in memory before writing
(can be faster for multi-file input at the cost of memory).}

\item{input_format}{Format of the input file(s), if it can't be
detected from the file extension.}

\item{overwrite}{If TRUE, allow overwriting an existing output file.}

\item{fsync}{If TRUE, call fsync after writing the output file.}

\item{generator}{Generator string to record in the output file header.}

\item{output_header}{Character vector of extra `key=value` header
options to set on the output file.}

\item{verbose}{If TRUE, print verbose progress information.}
}
\value{
If `output` is NULL, the written data as a character string.
  Otherwise, the `output` path, invisibly.
}
\description{
Wraps `osmium cat`.
}
