% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{osmium_merge}
\alias{osmium_merge}
\title{Merge several sorted OSM files into one}
\usage{
osmium_merge(
  input,
  output = NULL,
  with_history = FALSE,
  input_format = NULL,
  overwrite = FALSE,
  fsync = FALSE,
  generator = NULL,
  output_header = NULL,
  verbose = FALSE
)
}
\arguments{
\item{input}{Character vector of two or more input OSM file paths.}

\item{output}{Output file path.}

\item{with_history}{If TRUE, suppress the warning about input files
containing multiple versions of the same object.}

\item{input_format, overwrite, fsync, generator, output_header, verbose}{See [osmium_cat()].}
}
\value{
The `output` path, invisibly.
}
\description{
Wraps `osmium merge`. Input files must each be sorted by type then ID
(as `osmium sort` produces).
}
