% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{osmiumr_registered_commands}
\alias{osmiumr_registered_commands}
\title{List commands registered with the vendored osmium-tool CommandFactory}
\usage{
osmiumr_registered_commands()
}
\value{
character vector of command names
}
\description{
List commands registered with the vendored osmium-tool CommandFactory
}
