% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosmium-package.R
\docType{package}
\name{rosmium-package}
\alias{rosmium}
\alias{rosmium-package}
\title{rosmium: R Bindings for the 'osmium-tool' OpenStreetMap Command Line Tool}
\description{
Provides R functions wrapping the command implementations of 'osmium-tool' (\url{https://osmcode.org/osmium-tool/}), a command line tool for working with OpenStreetMap data built on the 'libosmium' library.
}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

Authors:
\itemize{
  \item Mark Padgham \email{mark.padgham@email.com}
  \item Jochen Topf \email{jochen@topf.org} (Author of the vendored osmium-tool C++ sources) [copyright holder]
}

}
\keyword{internal}
